Write make_move game logic
Will pattern match on game state. If game state is won or lost then the
game state does not change. So will return {game, tally} - tally will
have the same information as the game struct but the used letters will
actually be a list instead of a MapSet.

Write a single test to check if the state does not change for game won
or lost - will use a for loop to iterate over `[:won, :lost]` to run the
same code and assertions.