Write game logic for scoring the guesses
Guess gets added to used list. Then the guess is checked against the
letters in the word. If all the guesses match the letters then the game
is won if not a subset but still a good guess changes the state to
:good_guess.

If a bad guess `score_guess` needs to check if there is only one turn
left, if so game is lost. If more than one turn then update the state to
:bad_guess and decrement a turn.

Start writing a test to checks muliple guesses in a row and passing each
to `make_move`.