Create application for dictionary that uses Application behavior to start the agent process.
Add this to the application list in `mix.exs` in order for the process
to be started at runtime - requires a start function that will be
called, invokes the `Runtime.Server.start_link` to start up the agent
process with a registered name. Then can remove the `start_link` in the
Dictionary API since is invoked at runtime.