Add in form for user to enter guess
Add new route controllers for new game and update to update the game
struct on the server. Both the new actions will redirect to the url
`hangman/current` which is a show action that gets the updated tally
from the game pid on the conn and displays the information.

This prevents user from reloading page and submittng the previous
request of the form. Prevents form from being submitted more than once.