Fix dangling process of game server.
Front end was not signling to game server to shutdown if not in use.
Set up a watcher on game server - process sets up a timeout. Every time
the game server receives a request, it resets the timer. But should the
timer expire, that means that no requests have been received in
awhile and so the watchdog will simply shutdown the game server.