Start work on word so far component.
Add in game state, use a helper function to display a sentence instead
of the state itself. Also display each letter from tally.letters (either
a letter will be displayed if guess is good or will be an underscore).

Add a `phx-window-keyup` handle event to the game liveview div. Will
invoke `make_move` event if any key is pressed. Will pass that key as
the guess to `Hangman.make_move` and then the socket will be updated
with the new tally after each event.